# backend/docs/swagger.yaml
basePath: /
definitions:
...
  domain.DNSRecord:
    properties:
      createPtr:
        type: boolean
      id:
        type: string
      lastSyncedAt:
        type: string
      name:
        type: string
      ttl:
        type: integer
      type:
        type: string
      updatedAt:
        type: string
      value:
        type: string
      zoneId:
        type: string
    type: object
...
  router.dnsZoneCreateResponseDoc:
    properties:
      dynamicUpdate:
        type: string
      id:
        type: string
      lastError:
        type: string
      lastSyncedAt:
        type: string
      name:
        type: string
      records:
        items:
          $ref: '#/definitions/domain.DNSRecord'
        type: array
      reverse:
        type: boolean
      serverId:
        type: string
      syncStatus:
        type: string
      type:
        type: string
      warning:
        type: string
    type: object
  router.dnsZoneResponse:
    properties:
      dynamicUpdate:
        type: string
      id:
        type: string
      lastError:
        type: string
      lastSyncedAt:
        type: string
      name:
        type: string
      reverse:
        type: boolean
      serverId:
        type: string
      syncStatus:
        type: string
      type:
        type: string
    type: object
  router.errorDocResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  router.serverResponse:
    properties:
      agentUrl:
        type: string
      apiKey:
        type: string
      failureCount:
        type: integer
      host:
        type: string
      id:
        type: string
      lastChecked:
        type: string
      name:
        type: string
      role:
        type: string
      status:
        type: string
      tlsInsecure:
        type: boolean
    type: object
...
  router.statusResponse:
    properties:
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: ZoneLease Windows DNS / DHCP 统一管理控制台后端 API，提供认证、服务器登记、DNS/DHCP 资源管理、刷新事件和审计查询接口。
  title: ZoneLease API
  version: "1.0"
paths:
...
  /api/dns/zones:
    post:
      consumes:
      - application/json
      description: Agent 确认区域存在后，后端会立即读取该区域默认 SOA、NS 等记录并通过 records 返回；即时读取失败时区域仍创建成功，并通过
        warning 返回提示。
      parameters:
      - description: DNS 区域参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.dnsZoneResponse'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/router.dnsZoneCreateResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.errorDocResponse'
      security:
      - BearerAuth: []
      summary: 创建 DNS 区域
      tags:
      - DNS
...
  /api/servers:
    post:
      consumes:
      - application/json
      description: 保存前会检查 Agent 名称、接口地址是否重复和角色是否合法；Agent 连通性由前端先调用 /api/servers/probe
        完成，保存接口不再重复访问 Agent。
      parameters:
      - description: 服务器参数，包含 name、role、agentUrl、可选 apiKey 和可选 tlsInsecure
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.serverResponse'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/router.serverResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.errorDocResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.errorDocResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/router.errorDocResponse'
      security:
      - BearerAuth: []
      summary: 添加服务器
      tags:
      - Servers
  /api/servers/{id}:
    delete:
      parameters:
      - description: 服务器 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.statusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.errorDocResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.errorDocResponse'
      security:
      - BearerAuth: []
      summary: 删除服务器
      tags:
      - Servers
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"